+++ src/eval.hpp
@@ -33,6 +33,6 @@ using std::string;
 
 
 int material(Board);
-int material_weight(int);
+float material_weight(int);
 
-int eval(Board);
+float eval(Board);
